requirements.txt
Brotli==1.0.9
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dash==1.20.0
dash-bootstrap-components==0.12.2
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
memory-tempfile==2.2.3
numpy==1.20.3
pandas==1.2.4
Pillow==8.3.1
plotly==5.0.0
psycopg2-binary==2.9.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.18.0
pytz==2021.1
regex==2021.8.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.18
tenacity==7.0.0
tqdm==4.62.0
urllib3==1.26.6
Werkzeug==2.0.1
